# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SaveResPizzas": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"DollarPrice": {
						"type": "number",
						"format": "double"
					},
					"EuroPrice": {
						"type": "number",
						"format": "double"
					},
					"DollarDelivery": {
						"type": "number",
						"format": "double"
					},
					"EuroDelivery": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"Name",
					"DollarPrice",
					"EuroPrice",
					"DollarDelivery",
					"EuroDelivery"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqPizza": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"DollarPrice": {
						"type": "number",
						"format": "double"
					},
					"EuroPrice": {
						"type": "number",
						"format": "double"
					},
					"DollarDelivery": {
						"type": "number",
						"format": "double"
					},
					"EuroDelivery": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"Name",
					"DollarPrice",
					"EuroPrice",
					"DollarDelivery",
					"EuroDelivery"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResUser": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Cell": {
						"type": "string"
					},
					"Address": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"Name",
					"Email",
					"Cell",
					"Address"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqUser": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Cell": {
						"type": "string"
					},
					"Address": {
						"type": "string"
					}
				},
				"required": [
					"Name",
					"Email",
					"Cell",
					"Address"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqUser": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Cell": {
						"type": "string"
					},
					"Address": {
						"type": "string"
					}
				},
				"required": [
					"Name",
					"Email",
					"Cell",
					"Address"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConfirmResUser": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"ConfirmationCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"ConfirmationCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ConfirmReqUser": {
				"properties": {
					"Cell": {
						"type": "string"
					}
				},
				"required": [
					"Cell"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"USER": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"AuthReqUser": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"ConfirmationCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"ConfirmationCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CustomResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"age": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"age"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PizzaInterface": {
				"properties": {
					"Pizza": {
						"type": "string"
					},
					"Quantity": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"Pizza",
					"Quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResOrder": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Pizza": {
						"items": {
							"$ref": "#/components/schemas/PizzaInterface"
						},
						"type": "array"
					},
					"User": {
						"type": "string"
					},
					"Bill": {
						"type": "number",
						"format": "double"
					},
					"DeliveryCost": {
						"type": "number",
						"format": "double"
					},
					"TotalBill": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"_id",
					"Pizza",
					"User",
					"Bill",
					"DeliveryCost",
					"TotalBill"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqOrder": {
				"properties": {
					"Pizza": {
						"items": {
							"$ref": "#/components/schemas/PizzaInterface"
						},
						"type": "array"
					},
					"User": {
						"type": "string"
					},
					"Currency": {
						"type": "string"
					}
				},
				"required": [
					"Pizza",
					"User"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetOrderList": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "Pizza-Delivery-App",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Zaheer Ud Din Babar"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/pizza/savePizza": {
			"post": {
				"operationId": "SavePizza",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveResPizzas"
								}
							}
						}
					}
				},
				"tags": [
					"pizza"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqPizza"
							}
						}
					}
				}
			}
		},
		"/user/saveUser": {
			"post": {
				"operationId": "SaveUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResUser"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqUser"
							}
						}
					}
				}
			}
		},
		"/user/updateUser": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResUser"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqUser"
							}
						}
					}
				}
			}
		},
		"/user/confirmUser": {
			"post": {
				"operationId": "ConfirmUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConfirmResUser"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ConfirmReqUser"
							}
						}
					}
				}
			}
		},
		"/user/authUser": {
			"post": {
				"operationId": "AuthUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/USER"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AuthReqUser"
							}
						}
					}
				}
			}
		},
		"/user/getData": {
			"get": {
				"operationId": "GetData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CustomResponse"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": []
			}
		},
		"/order/saveOrder": {
			"post": {
				"operationId": "SaveOrder",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResOrder"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqOrder"
							}
						}
					}
				}
			}
		},
		"/order/getOrderList": {
			"post": {
				"operationId": "GetOrderList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SaveUpdateResOrder"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"order"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetOrderList"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}